@@ -15,68 +15,68 @@
         /// <summary>
         /// Period parameters M
         /// </summary>
-        private const Int32 M = 397;
+        private const int M = 397;
 
         /// <summary>
         /// Period parameters N
         /// </summary>
-        private const Int32 N = 624;
+        private const int N = 624;
 
         /// <summary>
         /// constant vector a
         /// </summary>
-        private const UInt32 MatrixA = 0x9908b0df;
+        private const uint MatrixA = 0x9908b0df;
 
         /// <summary>
         /// most significant w-r bits
         /// </summary>
-        private const UInt32 UpperMask = 0x80000000;
+        private const uint UpperMask = 0x80000000;
 
         /// <summary>
         /// least significant r bits
         /// </summary>
-        private const UInt32 LowerMask = 0x7fffffff;
+        private const uint LowerMask = 0x7fffffff;
 
         /// <summary>
         /// Tempering parameter B
         /// </summary>
-        private const UInt32 TemperingMaskB = 0x9d2c5680;
+        private const uint TemperingMaskB = 0x9d2c5680;
 
         /// <summary>
         /// Tempering parameter C
         /// </summary>
-        private const UInt32 TemperingMaskC = 0xefc60000;
+        private const uint TemperingMaskC = 0xefc60000;
 
         /// <summary>
         /// the array for the state vector
         /// </summary>
-        private readonly UInt32[] _mt = new UInt32[N];
+        private readonly uint[] _mt = new uint[N];
 
-        private Int16 _mti;
+        private short _mti;
 
-        private static readonly UInt32[] _mag01 = { 0x0, MatrixA };
+        private static readonly uint[] _mag01 = { 0x0, MatrixA };
 
         /// <summary>
         /// 9007199254740991.0 is the maximum Double value which the 53 significand
         /// can hold when the exponent is 0.
         /// </summary>
-        private const Double FiftyThreeBitsOfOnes = 9007199254740991.0;
+        private const double FiftyThreeBitsOfOnes = 9007199254740991.0;
 
         /// <summary>
         /// Multiply by inverse to (vainly?) try to avoid a division.
         /// </summary>
-        private const Double Inverse53BitsOfOnes = 1.0 / FiftyThreeBitsOfOnes;
+        private const double Inverse53BitsOfOnes = 1.0 / FiftyThreeBitsOfOnes;
 
-        private const Double OnePlus53BitsOfOnes = FiftyThreeBitsOfOnes + 1;
-        private const Double InverseOnePlus53BitsOfOnes = 1.0 / OnePlus53BitsOfOnes;
+        private const double OnePlus53BitsOfOnes = FiftyThreeBitsOfOnes + 1;
+        private const double InverseOnePlus53BitsOfOnes = 1.0 / OnePlus53BitsOfOnes;
 
         /// <summary>
         /// Initializes a new instance of the MersenneTwister class with a given seed.
         /// </summary>
         /// <param name="seed">A value to use as a seed.</param>
-        public MersenneTwister(Int32 seed)
+        public MersenneTwister(int seed)
         {
-            init((UInt32)seed);
+            init((uint)seed);
         }
 
         /// <summary>
@@ -94,106 +94,106 @@
         /// Initializes a new instance of the MersenneTwister class with a given seed vector.
         /// </summary>
         /// <param name="initKey">The array for initializing keys.</param>
-        public MersenneTwister(IList<Int32> initKey)
+        public MersenneTwister(IList<int> initKey)
         {
             if (initKey == null)
             {
                 throw new ArgumentNullException("initKey");
             }
 
-            var initArray = new UInt32[initKey.Count];
+            var initArray = new uint[initKey.Count];
 
             for (var i = 0; i < initKey.Count; ++i)
             {
-                initArray[i] = (UInt32)initKey[i];
+                initArray[i] = (uint)initKey[i];
             }
 
             init(initArray);
         }
 
         /// <summary>
-        /// Returns the next pseudo-random <see cref="UInt32"/>.
+        /// Returns the next pseudo-random <see cref="uint"/>.
         /// </summary>
-        /// <returns>A pseudo-random <see cref="UInt32"/> value.</returns>
-        public virtual UInt32 NextUInt32()
+        /// <returns>A pseudo-random <see cref="uint"/> value.</returns>
+        public virtual uint NextUInt32()
         {
             return GenerateUInt32();
         }
 
         /// <summary>
-        /// Returns the next pseudo-random <see cref="UInt32"/> 
+        /// Returns the next pseudo-random <see cref="uint"/> 
         /// up to <paramref name="maxValue"/>.
         /// </summary>
         /// <param name="maxValue">
         /// The maximum value of the pseudo-random number to create.
         /// </param>
         /// <returns>
-        /// A pseudo-random <see cref="UInt32"/> value which is at most <paramref name="maxValue"/>.
+        /// A pseudo-random <see cref="uint"/> value which is at most <paramref name="maxValue"/>.
         /// </returns>
-        public virtual UInt32 NextUInt32(UInt32 maxValue)
+        public virtual uint NextUInt32(uint maxValue)
         {
-            return (UInt32)(GenerateUInt32() / ((Double)UInt32.MaxValue / maxValue));
+            return (uint)(GenerateUInt32() / ((double)uint.MaxValue / maxValue));
         }
 
         /// <summary>
-        /// Returns the next pseudo-random <see cref="UInt32"/> at least 
+        /// Returns the next pseudo-random <see cref="uint"/> at least 
         /// <paramref name="minValue"/> and up to <paramref name="maxValue"/>.
         /// </summary>
         /// <param name="minValue">The minimum value of the pseudo-random number to create.</param>
         /// <param name="maxValue">The maximum value of the pseudo-random number to create.</param>
         /// <returns>
-        /// A pseudo-random <see cref="UInt32"/> value which is at least 
+        /// A pseudo-random <see cref="uint"/> value which is at least 
         /// <paramref name="minValue"/> and at most <paramref name="maxValue"/>.
         /// </returns>
         /// <exception cref="ArgumentOutOfRangeException">
         /// If <c><paramref name="minValue"/> &gt;= <paramref name="maxValue"/></c>.
         /// </exception>
-        public virtual UInt32 NextUInt32(UInt32 minValue, UInt32 maxValue) /* throws ArgumentOutOfRangeException */
+        public virtual uint NextUInt32(uint minValue, uint maxValue) /* throws ArgumentOutOfRangeException */
         {
             if (minValue >= maxValue)
             {
                 throw new ArgumentOutOfRangeException();
             }
 
-            return (UInt32)((GenerateUInt32() / ((Double)UInt32.MaxValue / (maxValue - minValue))) + minValue);
+            return (uint)((GenerateUInt32() / ((double)uint.MaxValue / (maxValue - minValue))) + minValue);
         }
 
         /// <summary>
-        /// Returns the next pseudo-random <see cref="Int32"/>.
+        /// Returns the next pseudo-random <see cref="int"/>.
         /// </summary>
-        /// <returns>A pseudo-random <see cref="Int32"/> value.</returns>
-        public override Int32 Next()
+        /// <returns>A pseudo-random <see cref="int"/> value.</returns>
+        public override int Next()
         {
-            return Next(Int32.MaxValue);
+            return Next(int.MaxValue);
         }
 
         /// <summary>
-        /// Returns the next pseudo-random <see cref="Int32"/> up to <paramref name="maxValue"/>.
+        /// Returns the next pseudo-random <see cref="int"/> up to <paramref name="maxValue"/>.
         /// </summary>
         /// <param name="maxValue">The maximum value of the pseudo-random number to create.</param>
         /// <returns>
-        /// A pseudo-random <see cref="Int32"/> value which is at most <paramref name="maxValue"/>.
+        /// A pseudo-random <see cref="int"/> value which is at most <paramref name="maxValue"/>.
         /// </returns>
         /// <exception cref="ArgumentOutOfRangeException">
         /// When <paramref name="maxValue"/> &lt; 0.
         /// </exception>
-        public override Int32 Next(Int32 maxValue)
+        public override int Next(int maxValue)
         {
             if (maxValue <= 1)
             {
                 if (maxValue < 0)
                 {
-                    throw new ArgumentOutOfRangeException();
+                    throw new ArgumentOutOfRangeException(nameof(maxValue));
                 }
 
                 return 0;
             }
 
-            return (Int32)(NextDouble() * maxValue);
+            return (int)(NextDouble() * maxValue);
         }
 
         /// <summary>
-        /// Returns the next pseudo-random <see cref="Int32"/> 
+        /// Returns the next pseudo-random <see cref="int"/> 
         /// at least <paramref name="minValue"/> 
         /// and up to <paramref name="maxValue"/>.
         /// </summary>
@@ -204,7 +204,7 @@
         /// <exception cref="ArgumentOutOfRangeException">
         /// If <c><paramref name="minValue"/> &gt;= <paramref name="maxValue"/></c>.
         /// </exception>
-        public override Int32 Next(Int32 minValue, Int32 maxValue)
+        public override int Next(int minValue, int maxValue)
         {
             if (maxValue <= minValue)
             {
@@ -226,24 +226,24 @@
         /// <exception cref="ArgumentNullException">
         /// If <c><paramref name="buffer"/> == <see langword="null"/></c>.
         /// </exception>
-        public override void NextBytes(Byte[] buffer)
+        public override void NextBytes(byte[] buffer)
         {
             // [codekaizen: corrected this to check null before checking length.]
             if (buffer == null)
             {
-                throw new ArgumentNullException();
+                throw new ArgumentNullException(nameof(buffer));
             }
 
-            Int32 bufLen = buffer.Length;
+            int bufLen = buffer.Length;
 
-            for (Int32 idx = 0; idx < bufLen; ++idx)
+            for (int idx = 0; idx < bufLen; ++idx)
             {
-                buffer[idx] = (Byte)Next(256);
+                buffer[idx] = (byte)Next(256);
             }
         }
 
         /// <summary>
-        /// Returns the next pseudo-random <see cref="Double"/> value.
+        /// Returns the next pseudo-random <see cref="double"/> value.
         /// </summary>
         /// <returns>A pseudo-random Double floating point value.</returns>
         /// <remarks>
@@ -270,7 +270,7 @@
         /// </code>
         /// </para>
         /// </remarks>
-        public override Double NextDouble()
+        public override double NextDouble()
         {
             return compute53BitRandom(0, InverseOnePlus53BitsOfOnes);
         }
@@ -293,7 +293,7 @@
         /// returns a Double-precision pseudo-random number greater than or equal to zero and
         /// strictly less than one.
         /// </returns>
-        public Double NextDouble(Boolean includeOne)
+        public double NextDouble(bool includeOne)
         {
             return includeOne ? compute53BitRandom(0, Inverse53BitsOfOnes) : NextDouble();
         }
@@ -302,7 +302,7 @@
         /// Returns a pseudo-random number greater than 0.0 and less than 1.0.
         /// </summary>
         /// <returns>A pseudo-random number greater than 0.0 and less than 1.0.</returns>
-        public Double NextDoublePositive()
+        public double NextDoublePositive()
         {
             return compute53BitRandom(0.5, Inverse53BitsOfOnes);
         }
@@ -314,9 +314,9 @@
         /// A single-precision floating point number greater than or equal to 0.0, 
         /// and less than 1.0.
         /// </returns>
-        public Single NextSingle()
+        public float NextSingle()
         {
-            return (Single)NextDouble();
+            return (float)NextDouble();
         }
 
         /// <summary>
@@ -336,33 +336,33 @@
         /// this method returns a single-precision pseudo-random number greater than or equal to zero and
         /// strictly less than one.
         /// </returns>
-        public Single NextSingle(Boolean includeOne)
+        public float NextSingle(bool includeOne)
         {
-            return (Single)NextDouble(includeOne);
+            return (float)NextDouble(includeOne);
         }
 
         /// <summary>
         /// Returns a pseudo-random number greater than 0.0 and less than 1.0.
         /// </summary>
         /// <returns>A pseudo-random number greater than 0.0 and less than 1.0.</returns>
-        public Single NextSinglePositive()
+        public float NextSinglePositive()
         {
-            return (Single)NextDoublePositive();
+            return (float)NextDoublePositive();
         }
 
         /// <summary>
-        /// Generates a new pseudo-random <see cref="UInt32"/>.
+        /// Generates a new pseudo-random <see cref="uint"/>.
         /// </summary>
-        /// <returns>A pseudo-random <see cref="UInt32"/>.</returns>
-        public UInt32 GenerateUInt32()
+        /// <returns>A pseudo-random <see cref="uint"/>.</returns>
+        public uint GenerateUInt32()
         {
-            UInt32 y;
+            uint y;
 
             /* _mag01[x] = x * MatrixA  for x=0,1 */
             if (_mti >= N)
             {
                 /* generate N words at one time */
-                Int16 kk = 0;
+                short kk = 0;
 
                 for (; kk < N - M; ++kk)
                 {
@@ -391,27 +391,27 @@
             return y;
         }
 
-        private static UInt32 temperingShiftU(UInt32 y)
+        private static uint temperingShiftU(uint y)
         {
             return (y >> 11);
         }
 
-        private static UInt32 temperingShiftS(UInt32 y)
+        private static uint temperingShiftS(uint y)
         {
             return (y << 7);
         }
 
-        private static UInt32 temperingShiftT(UInt32 y)
+        private static uint temperingShiftT(uint y)
         {
             return (y << 15);
         }
 
-        private static UInt32 temperingShiftL(UInt32 y)
+        private static uint temperingShiftL(uint y)
         {
             return (y >> 18);
         }
 
-        private void init(UInt32 seed)
+        private void init(uint seed)
         {
             _mt[0] = seed & 0xffffffffU;
 
@@ -427,12 +427,12 @@
             }
         }
 
-        private void init(UInt32[] key)
+        private void init(uint[] key)
         {
-            Int32 i, j, k;
+            int i, j, k;
             init(19650218U);
 
-            Int32 keyLength = key.Length;
+            int keyLength = key.Length;
             i = 1; j = 0;
             k = (N > keyLength ? N : keyLength);
 
@@ -462,12 +462,12 @@
             _mt[0] = 0x80000000U; // MSB is 1; assuring non-zero initial array
         }
 
-        private Double compute53BitRandom(Double translate, Double scale)
+        private double compute53BitRandom(double translate, double scale)
         {
             // get 27 pseudo-random bits
-            UInt64 a = (UInt64)GenerateUInt32() >> 5;
+            ulong a = (ulong)GenerateUInt32() >> 5;
             // get 26 pseudo-random bits
-            UInt64 b = (UInt64)GenerateUInt32() >> 6;
+            ulong b = (ulong)GenerateUInt32() >> 6;
 
             // shift the 27 pseudo-random bits (a) over by 26 bits (* 67108864.0) and
             // add another pseudo-random 26 bits (+ b).
